Labfiles/02-build-ai-agent/Python/requirements.txt
# requirements.txt
openai
python-dotenv
pandas
matplotlib
ipython
